Moved all but the two answer tests to test_particle_trajectories_pytest.py and removed the nose versions of test_default_field_tuple and test_time_and_index
