+++ nose_ignores.txt
@@ -36,3 +36,4 @@
 --ignore-file=test_version\.py
 --ignore-file=test_gadget_pytest\.py
 --ignore-file=test_vr_orientation\.py
+--ignore-file=test_particle_trajectories_pytest.py
